Rename `ValkeyAdapter` package to `variables` and update relevant imports
