+++ .golangci.yml
@@ -24,6 +24,8 @@ linters:
     revive:
       rules:
         - name: comment-spacings
+        - name: import-shadowing
+        - name: redundant-import-alias
   exclusions:
     generated: lax
     rules:
