+++ kube/kubetest/fake_store.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/mydecisive/mdai-data-core/kube"
-	v1 "k8s.io/api/core/v1"
+	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
