+++ kube/secret_informer.go
@@ -11,7 +11,7 @@ import (
 	corev1 "k8s.io/client-go/listers/core/v1"
 
 	"go.uber.org/zap"
-	v1 "k8s.io/api/core/v1"
+	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
